don't use SequenceCommand when conflating multiple matches
ReplaceGeometry currently changes relations in such a way that we can't string commands together in one SequenceCommand, as only the last match will be done correctly as far as relations go. Unfortunately for a large list of matches this won't be very tidy as far as the command stack goes.